% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Bag of Little Boostraps Linear Regression}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE, cl = NULL)
}
\arguments{
\item{formula}{formula}

\item{data}{dataset of interest}

\item{m}{int: number of groups to split data into, used in bootstrapping process}

\item{B}{int: number of bootstrap samples}

\item{parallel}{boolean: defines whether or not to use parallelization. Note: if TRUE, run furrr::plan(multisession, worker = desired_number_of_CPUs) in console before using function.}

\item{cl}{int: desired number of clusters}
}
\value{
linear regression model
}
\description{
Bag of Little Boostraps Linear Regression
}
