% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{coef.blb_logreg}
\alias{coef.blb_logreg}
\title{Coefficients}
\usage{
\method{coef}{blb_logreg}(object, ...)
}
\arguments{
\item{object}{blb_logreg object}

\item{...}{unused argments}
}
\value{
coefficients of parameters in logistic regression model
}
\description{
Coefficients
}
