% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Coefficients}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{blb_logreg object}

\item{...}{unused arguments}
}
\value{
coefficients of parameters in linear regression model
}
\description{
Coefficients
}
