% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{confint.blb_logreg}
\alias{confint.blb_logreg}
\title{Confidence Interval for Mean}
\usage{
\method{confint}{blb_logreg}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blb_logreg object}

\item{parm}{string/character vector: parameter(s) of interest}

\item{level}{numeric: confidence level}

\item{...}{unused arguments}
}
\value{
confidence interval of parameters of interest, all parameters by default
}
\description{
Confidence Interval for Mean
}
