% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Confidence interval of Mean}
\usage{
\method{confint}{blblm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{blblm object}

\item{parm}{string/character vector: parameter(s) of interest}

\item{level}{numeric: confidence level}

\item{...}{unused arguments}
}
\value{
confidence interval of parameters of interest, all parameters by default
}
\description{
Confidence interval of Mean
}
