% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{predict.blb_logreg}
\alias{predict.blb_logreg}
\title{Predict}
\usage{
\method{predict}{blb_logreg}(object, new_data, p = FALSE, ...)
}
\arguments{
\item{object}{blb_logreg object}

\item{new_data}{data set of interest}

\item{p}{boolean: probability; if TRUE, returns P(Y = 1|X = x), instead of binary {0,1} prediction}

\item{...}{unused arguments}
}
\value{
vector of predicted values of response variable
}
\description{
Predict
}
