% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Predict}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blblm object}

\item{new_data}{data set of interest}

\item{confidence}{boolean: confidence interval for predicted value}

\item{level}{numeric: confidence level}

\item{...}{unused arguments}
}
\value{
vector of predicted values of response variable
}
\description{
Predict
}
