% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblogreg.R
\name{print.blb_logreg}
\alias{print.blb_logreg}
\title{Print}
\usage{
\method{print}{blb_logreg}(x, ...)
}
\arguments{
\item{x}{blb_logreg object}

\item{...}{unused arguments}
}
\value{
prints model
}
\description{
Print
}
