% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Standard Deviation}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{blb_logreg object}

\item{confidence}{boolean: confidence interval}

\item{level}{int: confidence level}

\item{...}{unused arguments}
}
\value{
standard deviation
}
\description{
Standard Deviation
}
